Import fixes to unbreak main use cases.
